# fix #293203: Cut down preferences save time to 1/3 for most operations + protect against multiple clicks which make the process even slower