--- a/mscore/musescore.cpp
+++ b/mscore/musescore.cpp
@@ -448,7 +448,7 @@ void updateExternalValuesFromPreferences() {
 //   preferencesChanged
 //---------------------------------------------------------
 
-void MuseScore::preferencesChanged(bool fromWorkspace)
+void MuseScore::preferencesChanged(bool fromWorkspace, bool changeUI)
       {
       updateExternalValuesFromPreferences();
 
@@ -460,7 +460,8 @@ void MuseScore::preferencesChanged(bool fromWorkspace)
       getAction("show-tours")->setChecked(preferences.getBool(PREF_UI_APP_STARTUP_SHOWTOURS));
       _statusBar->setVisible(preferences.getBool(PREF_UI_APP_SHOWSTATUSBAR));
 
-      MuseScore::updateUiStyleAndTheme();
+      if (changeUI)
+            MuseScore::updateUiStyleAndTheme(); // this is a slow operation
       updateIcons();
 
       QString fgWallpaper = preferences.getString(PREF_UI_CANVAS_FG_WALLPAPER);