--- a/mscore/prefsdialog.h
+++ b/mscore/prefsdialog.h
@@ -39,13 +39,30 @@ class PreferenceDialog : public AbstractDialog, private Ui::PrefsDialogBase {
       QMap<QString, Shortcut*> localShortcuts;
       bool shortcutsChanged;
       QButtonGroup* recordButtons;
+      std::vector<PreferenceItem*> normalWidgets;
+      std::vector<PreferenceItem*> uiRelatedWidgets;
+      std::vector<PreferenceItem*> audioRelatedWidgets;
+      std::vector<PreferenceItem*> modifiedWidgets;
+      std::vector<PreferenceItem*> modifiedUiWidgets;
+      std::vector<PreferenceItem*> modifiedAudioWidgets;
       PreferencesListWidget* advancedWidget;
 
       virtual void hideEvent(QHideEvent*);
       void apply();
       void updateSCListView();
       void setUseMidiOutput(bool);
-      void updateValues(bool useDefaultValues = false);
+      void updateValues(bool useDefaultValues = false, bool setup = false);
+      void checkApplyActivation();
+
+      void applySetActive(bool active);
+      void updateShortestNote();
+      void applyShortestNote();
+      void languageUpdate();
+      void languageApply();
+      void updateCharsetListGP();
+      void updateCharsetListOve();
+      void updateUseLocalAvsOmr();
+      void applyPageVertical();
 
    private slots:
       void buttonBoxClicked(QAbstractButton*);
@@ -82,13 +99,16 @@ class PreferenceDialog : public AbstractDialog, private Ui::PrefsDialogBase {
       void filterAdvancedPreferences(const QString&);
       void resetAdvancedPreferenceToDefault();
       void restartAudioEngine();
+      void widgetModified();
+      void uiWidgetModified();
+      void audioWidgetModified();
+      void applyActivate();
 
       void changeSoundfontPaths();
       void updateTranslationClicked();
 
    signals:
-      void preferencesChanged();
-      void preferencesChangedWithBool(bool b);
+      void preferencesChanged(bool fromWorkspace, bool changeUI);
       void mixerPreferencesChanged(bool showMidiControls);
 
    protected: