… mscore/preferenceslistwidget.cpp
… mscore/prefsdialog.cpp