Resolves: https://musescore.org/en/node/293203

Applying preference changes is pretty slow (even if no changes are made).

The 2 main reasons behind the slow down are the UI/STYLE update in PreferencesChanged and the fact that PreferencesChanged was called twice. I disabled the second call and disabled the UI/STYLE update when no ui related preferences are changed. 

This leads to a pretty significant speedup:
1s -> 300ms for general changes (or no changes)
1s -> 600ms for ui related changes

Further improvements can be made if:

`mscore->changeWorkspace(WorkspacesManager::currentWorkspace(), true);`

is disabled. I tried disabling it and things seemed normal, but I don't know why it is there, so I let it be.

(time counted using QElapsedTimer on Linux, Ubuntu 19.10)

- [x] I signed [CLA](https://musescore.org/en/cla)
- [x] I made sure the code in the PR follows [the coding rules](https://musescore.org/en/handbook/developers-handbook/finding-your-way-around/musescore-coding-rules)
- [x] I made sure the code compiles on my machine
- [x] I made sure there are no unnecessary changes in the code
- [x] I made sure the title of the PR reflects the core meaning of the issue you are solving
- [x] I made sure the commit message(s) contain a description and answer the question "Why do those changes fix that particular issue?" or "Why are those changes really necessary as improvements?"
- [x] I made sure the commit message title starts with "fix #424242:" if there is a related issue
- [n\a] I created the test (mtest, vtest, script test) to verify the changes I made