@MarcSabatella 2020-04-10
In case it helps to understand the background of the relationship between workspaces and preferences:

Originally, workspaces were only about palette and toolbar configuration.  GUI settings like font, icon size, etc. were always global.  As part of a GSoC project a couple of years ago, this was expanded to allow GUI settings to optionally be included as part of a workspace.  So when you create a workspace - or use View / Workspaces / Edit on an existing one - there is a checkbox where you can say you want GUI settings to be included within the workspace rather than globally.  Thus you can create a workspace that uses bigger fonts and icons, while keeping default fonts in other workspaces.  I have one I created for educational purposes that switches the interface to Comic Sans (!) in large sizes, to look more "kid friendly".

So every time you switch workspaces, those preferences potentially get updated.  Conversely, every time you make a change to preferences, the workspace needs to be informed so it can see if it needs to updates its own settings (if the workspace was configured to include GUI preferences).  And when saving preferences, the code needs to know whether the preference should be saved globally (like if I change my font to Comic Sans while in the Basic workspace) or only within the workspace (like when I changed it within my "kid friendly" workspace that has GUI preferences enabled).

No doubt it's bad that code is being executed twice unnecessarily, so I'm glad you're trying to streamline this.  But maybe that helps understand where these different paths come from and how you can test that your changes don't break any of this.