@SKefalidis 2020-04-23
So, this PR became a bit more ambitious. I am essentially trying to treat all preferences like the advanced preferences. This should reduce the effort of adding new preferences, it should increase the speed of applying changes (since only the modified preferences will be saved) and it allows me to disable 'Apply' when nothing has changed.

WIP