@SKefalidis 2020-04-25
Updated the PR (see the OP). Most of it is done and most preferences work correctly. I await opinions :-)

@MarcSabatella thank you very much for the workspace info (dodged a brown bag thanks to that :-)