@SKefalidis 2020-04-27
I seem to have some trouble rebasing, I'll deal with that tomorrow. Most of the proposed changes have been made.