@Marr11317 2020-04-28
the cheat sheet: 
```
git fetch --all && git checkout master && git rebase upstream/master
git checkout your-branch && git rebase master
```
You'll then have to resolve the conflicts. When they are fixed, run
'''
git rebase --continue
'''
At that point, you'll have all your commits added over the current master
'''
git rebase -i #last-commit-in-master-hash
'''
Take oddly, squashing goes from bottom to top (commits accumulate in piles, apparently ;))
so you want to take all these commits after the first one, and replace `pick` with `squash`.

There is probably a faster way (like squashing your changes with -i with the first rebase, but I'd say that if you want to understand how rebasing really works, you should first separate this in two actions: first rebasing over master, then squashing your commits.

I hope this helps, two ago it was @Jojo-Schmitz that gave me all this info, and it helped me stop from doing these frustrating errors.
Thanks again @Jojo-Schmitz