@Jojo-Schmitz 2020-04-28
to rebase I use `git pull --rebase musescore master` (with `musescore` being the name of my upstream repo), rix the conflicts, `git add` the fixed file(s), `git rebase --continue` and finally `git push -f`. To squash though follow the above steps.