@SKefalidis 2020-04-28
Hm.. appveyor is giving an error, but github says (and I agree) that there are no conflicts.