@SKefalidis 2020-04-28
No longer WIP. There is one more addition that I would like to make, but I would preferably do that in another PR (added a TODO comment). 
I've tested most preferences and they seem to work (apply/reset/update). I was unable to test a couple of preferences in the IO tab since they are missing or crashing (as they do in the master branch for me).