@SKefalidis 2020-04-28
Fixed a crash caused by moving the widget vectors to the constructor (from PreferenceDialog::start).