@SKefalidis 2020-05-01
Added a check that uses `isModified` to determine if the apply button should be active. If a user changes a preference and then reverts the change, apply will be disabled again (if there are no other changes).