@SKefalidis 2020-05-02
Added comments, deleted placeholders, repositioned some PreferenceItems.